feat: added validation rules for prescription store

+++ database/migrations/2023_01_10_172430_create_prescriptions_table.php
@@ -21,9 +21,9 @@ return new class extends Migration
             $table->char('problem_id', 36)->nullable();
             $table->char('user_id', 36)->nullable();
             $table->date('prescription_date');
-            $table->bigInteger('quantity')->unsigned();
+            $table->smallInteger('quantity')->unsigned();
             $table->string('dosage');
-            $table->bigInteger('duration')->unsigned()->nullable();
+            $table->smallInteger('duration')->unsigned()->nullable();
             $table->boolean('in_evidence');
             $table->text('notes')->nullable();
             $table->boolean('print_notes')->default(false);
